requirements.txt
slack_bolt
snowflake
snowflake-snowpark-python
requests
pandas
numpy
python-dotenv